+++ CHANGES.md
@@ -1,4 +1,7 @@
 <!-- Unreleased: append new entries here -->
 
-* [!33](https://github.com/tezos-commons/homebase-lite/pull/33)
+* [#38](https://github.com/tezos-commons/homebase-lite/pull/38)
+  Update morley
+  + Capitalization in the storage field `FA2Config` changed to `fa2Config`
+* [#33](https://github.com/tezos-commons/homebase-lite/pull/33)
   Rename `senderIsNotAdmin_` to `senderIsNotAdmin`